add tab completion for run command
