+++ lib/rex/ui/text/dispatcher_shell.rb
@@ -397,9 +397,10 @@ module DispatcherShell
     #
     def tab_complete_option_values(str, words, opt:)
       # Readline.completion_append_character = " "
+
       res = []
       mod = active_module
-      # require "pry"; binding.pry
+
       # With no active module, we have nothing to compare
       if (not mod)
         return res
@@ -432,9 +433,13 @@ module DispatcherShell
         return option_values_sessions() if opt.upcase == 'SESSION'
       end
 
+      # require "pry"; binding.pry
       # Is this option used by the active module?
       mod.options.each_key do |key|
-        res.concat(option_values_dispatch(mod.options[key], str, words)) if key.downcase == opt.downcase
+        if key.downcase == opt.downcase
+          # require "pry"; binding.pry
+          res.concat(option_values_dispatch(mod.options[key], str, words))
+        end
       end
 
       # How about the selected payload?
@@ -449,6 +454,199 @@ module DispatcherShell
       return res
     end
 
+    #
+    # Provide possible option values based on type
+    #
+    def option_values_dispatch(o, str, words)
+      # require "pry"; binding.pry
+      res = []
+      res << o.default.to_s if o.default
+
+      case o
+      when Msf::OptAddress
+        case o.name.upcase
+        when 'RHOST'
+          option_values_target_addrs().each do |addr|
+            res << addr
+          end
+        when 'LHOST', 'SRVHOST', 'REVERSELISTENERBINDADDRESS'
+          rh = self.active_module.datastore['RHOST'] || framework.datastore['RHOST']
+          if rh and not rh.empty?
+            res << Rex::Socket.source_address(rh)
+          else
+            res += tab_complete_source_address
+            res += tab_complete_source_interface(o)
+          end
+        end
+
+      when Msf::OptAddressRange
+        case str
+        when /^file:(.*)/
+          files = tab_complete_filenames($1, words)
+          res += files.map { |f| "file:" + f } if files
+        when /\/$/
+          res << str+'32'
+          res << str+'24'
+          res << str+'16'
+        when /\-$/
+          res << str+str[0, str.length - 1]
+        else
+          option_values_target_addrs().each do |addr|
+            res << addr
+          end
+        end
+
+      when Msf::OptPort
+        case o.name.upcase
+        when 'RPORT'
+          option_values_target_ports().each do |port|
+            res << port
+          end
+        end
+
+        if (res.empty?)
+          res << (rand(65534)+1).to_s
+        end
+
+      when Msf::OptEnum
+        o.enums.each do |val|
+          res << val
+        end
+
+      when Msf::OptPath
+        files = tab_complete_filenames(str, words)
+        res += files if files
+
+      when Msf::OptBool
+        res << 'true'
+        res << 'false'
+
+      when Msf::OptString
+        if (str =~ /^file:(.*)/)
+          files = tab_complete_filenames($1, words)
+          res += files.map { |f| "file:" + f } if files
+        end
+      end
+      require "pry"; binding.pry
+      return res
+    end
+
+    #
+    # Provide valid payload options for the current exploit
+    #
+    def option_values_payloads
+      if @cache_payloads && active_module == @previous_module && active_module.target == @previous_target
+        return @cache_payloads
+      end
+
+      @previous_module = active_module
+      @previous_target = active_module.target
+
+      @cache_payloads = active_module.compatible_payloads.map do |refname, payload|
+        refname
+      end
+
+      @cache_payloads
+    end
+
+    #
+    # Provide valid session options for the current post-exploit module
+    #
+    def option_values_sessions
+      if active_module.respond_to?(:compatible_sessions)
+        active_module.compatible_sessions.map { |sid| sid.to_s }
+      end
+    end
+
+    #
+    # Provide valid target options for the current exploit
+    #
+    def option_values_targets
+      res = []
+      if (active_module.targets)
+        1.upto(active_module.targets.length) { |i| res << (i-1).to_s }
+        res += active_module.targets.map(&:name)
+      end
+      return res
+    end
+
+
+    #
+    # Provide valid action options for the current module
+    #
+    def option_values_actions
+      res = []
+      if (active_module.actions)
+        active_module.actions.each { |i| res << i.name }
+      end
+      return res
+    end
+
+    #
+    # Provide valid nops options for the current exploit
+    #
+    def option_values_nops
+      framework.nops.map { |refname, mod| refname }
+    end
+
+    #
+    # Provide valid encoders options for the current exploit or payload
+    #
+    def option_values_encoders
+      framework.encoders.map { |refname, mod| refname }
+    end
+
+    #
+    # Provide the target addresses
+    #
+    def option_values_target_addrs
+      res = [ ]
+      res << Rex::Socket.source_address()
+      return res if not framework.db.active
+
+      # List only those hosts with matching open ports?
+      mport = self.active_module.datastore['RPORT']
+      if mport
+        mport = mport.to_i
+        hosts = {}
+        framework.db.services.each do |service|
+          if service.port == mport
+            hosts[ service.host.address ] = true
+          end
+        end
+
+        hosts.keys.each do |host|
+          res << host
+        end
+
+        # List all hosts in the database
+      else
+        framework.db.hosts.each do |host|
+          res << host.address
+        end
+      end
+
+      return res
+    end
+
+    #
+    # Provide the target ports
+    #
+    def option_values_target_ports
+      res = [ ]
+      return res if not framework.db.active
+      return res if not self.active_module.datastore['RHOST']
+      host = framework.db.has_host?(framework.db.workspace, self.active_module.datastore['RHOST'])
+      return res if not host
+
+      framework.db.services.each do |service|
+        if service.host_id == host.id
+          res << service.port.to_s
+        end
+      end
+
+      return res
+    end
 
     #
     # Return a list of possible source addresses for tab completion.
@@ -571,9 +769,6 @@ module DispatcherShell
   def tab_complete_helper(dispatcher, str, words)
     items = []
 
-    require 'pry'; binding.pry
-    # ::Readline.completion_append_character = " "
-    require 'pry'; binding.pry
     tabs_meth = "cmd_#{words[0]}_tabs"
     # Is the user trying to tab complete one of our commands?
     if (dispatcher.commands.include?(words[0]) and dispatcher.respond_to?(tabs_meth))
@@ -585,7 +780,6 @@ module DispatcherShell
       return []
     end
 
-    ::Readline.completion_append_character ||= ' '
     return items
   end
